First batch of unit testing and some class changes
Made first unit test of constructor and an in progress test of the check
overdue unit test involving multiple classes, which was giving errors
(document yet to be produced due to time constraints at time of
writing). Also some tweaks to Loan class due to mistakes with two
functions not being properly implemented as the system operation
document had specified.